+++ src/Container/Definition/SelfDefinition.php
@@ -7,7 +7,7 @@ class SelfDefinition extends AbstractDefinition
 {
     public function __construct(string $className)
     {
-        parent::__construct($className, str_replace("\\", "__", $className), "");
+        parent::__construct($className, "");
     }
 
     public function needsDependencyResolution(): bool